# api/openapi.yaml
openapi: 3.1.0
info:
  title: omni-5e
  description: REST API for the D&D 5e System Reference Document
  version: 0.1.0
  license:
    name: CC-BY-4.0
    url: https://creativecommons.org/licenses/by/4.0/legalcode
  contact:
    name: omni-5e
    url: https://github.com/omni-5e/omni-5e
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
paths:
...
  /species:
    get:
      summary: List species (races)
      operationId: listSpecies
      tags: [species]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: size
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeciesList"
  /species/{slug}:
    get:
      summary: Get a species by slug
      operationId: getSpecies
      tags: [species]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Species detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Species"
        "404":
          $ref: "#/components/responses/NotFound"
  /backgrounds:
    get:
      summary: List backgrounds
      operationId: listBackgrounds
      tags: [backgrounds]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated list of backgrounds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackgroundList"
  /backgrounds/{slug}:
    get:
      summary: Get a background by slug
      operationId: getBackground
      tags: [backgrounds]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Background detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Background"
        "404":
          $ref: "#/components/responses/NotFound"
  /feats:
    get:
      summary: List feats
      operationId: listFeats
      tags: [feats]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of feats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatList"
  /feats/{slug}:
    get:
      summary: Get a feat by slug
      operationId: getFeat
      tags: [feats]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Feat detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Feat"
        "404":
          $ref: "#/components/responses/NotFound"
  /equipment:
    get:
      summary: List equipment
      operationId: listEquipment
      tags: [equipment]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: category
          in: query
          schema: { type: string, enum: [weapon, armor, tool, gear, mount, vehicle] }
      responses:
        "200":
          description: Paginated list of equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentList"
  /equipment/{slug}:
    get:
      summary: Get equipment by slug
      operationId: getEquipment
      tags: [equipment]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Equipment detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Equipment"
        "404":
          $ref: "#/components/responses/NotFound"
  /magic-items:
    get:
      summary: List magic items
      operationId: listMagicItems
      tags: [magic-items]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: rarity
          in: query
          schema: { type: string }
        - name: attunement
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Paginated list of magic items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MagicItemList"
  /magic-items/{slug}:
    get:
      summary: Get a magic item by slug
      operationId: getMagicItem
      tags: [magic-items]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Magic item detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/MagicItem"
        "404":
          $ref: "#/components/responses/NotFound"
  /conditions:
    get:
      summary: List conditions
      operationId: listConditions
      tags: [conditions]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated list of conditions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionList"
  /conditions/{slug}:
    get:
      summary: Get a condition by slug
      operationId: getCondition
      tags: [conditions]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Condition detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Condition"
        "404":
          $ref: "#/components/responses/NotFound"
  /glossary:
    get:
      summary: List glossary terms
      operationId: listGlossaryTerms
      tags: [glossary]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of glossary terms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlossaryTermList"
  /glossary/{slug}:
    get:
      summary: Get a glossary term by slug
      operationId: getGlossaryTerm
      tags: [glossary]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Glossary term detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/GlossaryTerm"
        "404":
          $ref: "#/components/responses/NotFound"
  /rules:
    get:
      summary: List rule sections
      operationId: listRuleSections
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/srd_version"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: source_file
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated list of rule sections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleSectionList"
  /rules/{slug}:
    get:
      summary: Get a rule section by slug
      operationId: getRuleSection
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/slug"
        - $ref: "#/components/parameters/srd_version"
      responses:
        "200":
          description: Rule section detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/RuleSection"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    srd_version:
      name: srd_version
      in: query
      schema: { type: string }
      description: SRD content version (e.g. 5.2.1). Also via X-SRD-Version header.
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
    slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
  schemas:
    ProblemDetail:
      type: object
      properties:
        type: { type: string, example: "about:blank" }
        title: { type: string, example: "Not Found" }
        status: { type: integer, example: 404 }
        detail: { type: string }
...
    PaginationMeta:
      type: object
      properties:
        srd_version: { type: string }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    PaginationLinks:
      type: object
      properties:
        self: { type: string }
        next: { type: string, nullable: true }
        prev: { type: string, nullable: true }
...
    Species:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        size: { type: string }
        speed: { type: integer }
        traits: { type: array, items: { $ref: "#/components/schemas/Trait" } }
        description: { type: string }
    SpeciesList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Species" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    Trait:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
    Background:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        ability_score_options: { type: object }
        skill_proficiencies: { type: array, items: { type: string } }
        granted_feat_slug: { type: string }
        equipment: { type: array, items: { type: string } }
        description: { type: string }
    BackgroundList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Background" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    Feat:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        category: { type: string }
        prerequisite: { type: string }
        description: { type: string }
        repeatable: { type: boolean }
    FeatList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Feat" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    Equipment:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        category: { type: string }
        cost: { type: string }
        weight: { type: number }
        properties: { type: object }
        description: { type: string }
    EquipmentList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Equipment" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    MagicItem:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        rarity: { type: string }
        requires_attunement: { type: boolean }
        type: { type: string }
        description: { type: string }
    MagicItemList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/MagicItem" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    Condition:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        description: { type: string }
    ConditionList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Condition" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    GlossaryTerm:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        category: { type: string }
        definition: { type: string }
    GlossaryTermList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/GlossaryTerm" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    RuleSection:
      type: object
      properties:
        slug: { type: string }
        url: { type: string }
        srd_version: { type: string }
        name: { type: string }
        source_file: { type: string }
        heading_path: { type: array, items: { type: string } }
        body: { type: string }
    RuleSectionList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/RuleSection" } }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }